[SELC-4534] feat: added request mapping users v2 to UserV2Controller (#381)
