+++ web/src/main/java/it/pagopa/selfcare/dashboard/web/controller/UserV2Controller.java
@@ -7,15 +7,11 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import it.pagopa.selfcare.commons.base.logging.LogUtils;
-import it.pagopa.selfcare.commons.base.security.SelfCareUser;
 import it.pagopa.selfcare.commons.web.model.Problem;
-import it.pagopa.selfcare.dashboard.connector.model.institution.InstitutionBase;
 import it.pagopa.selfcare.dashboard.connector.model.user.User;
 import it.pagopa.selfcare.dashboard.core.UserV2Service;
-import it.pagopa.selfcare.dashboard.web.InstitutionBaseResource;
 import it.pagopa.selfcare.dashboard.web.model.SearchUserDto;
 import it.pagopa.selfcare.dashboard.web.model.UpdateUserDto;
-import it.pagopa.selfcare.dashboard.web.model.mapper.InstitutionResourceMapper;
 import it.pagopa.selfcare.dashboard.web.model.mapper.UserMapperV2;
 import it.pagopa.selfcare.dashboard.web.model.user.UserResource;
 import lombok.RequiredArgsConstructor;
@@ -23,41 +19,19 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.Collection;
-import java.util.List;
 
 @Slf4j
 @RestController
 @Api(tags = "user")
-@RequestMapping(value = "/v2", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(value = "/v2/users", produces = MediaType.APPLICATION_JSON_VALUE)
 @RequiredArgsConstructor
 public class UserV2Controller {
 
     private final UserV2Service userService;
-    private final InstitutionResourceMapper institutionResourceMapper;
     private final UserMapperV2 userMapperV2;
-
-    @GetMapping("/institutions")
-    @ResponseStatus(HttpStatus.OK)
-    @ApiOperation(value = "", notes = "${swagger.dashboard.institutions.api.getInstitutions}")
-    public List<InstitutionBaseResource> getInstitutions(Authentication authentication) {
-
-        log.trace("getInstitutions start");
-        String userId = ((SelfCareUser) authentication.getPrincipal()).getId();
-        Collection<InstitutionBase> institutions = userService.getInstitutions(userId);
-
-        List<InstitutionBaseResource> result = institutions.stream()
-                .map(institutionResourceMapper::toResource)
-                .toList();
-        log.debug(LogUtils.CONFIDENTIAL_MARKER, "getInstitutions result = {}", result);
-        log.trace("getInstitutions end");
-
-        return result;
-    }
 
     @PostMapping(value = "/{userId}/suspend")
     @ResponseStatus(HttpStatus.NO_CONTENT)
