+++ web/src/test/java/it/pagopa/selfcare/dashboard/web/controller/InstitutionV2ControllerTest.java
@@ -1,10 +1,15 @@
 package it.pagopa.selfcare.dashboard.web.controller;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import it.pagopa.selfcare.commons.base.security.SelfCareUser;
+import it.pagopa.selfcare.dashboard.connector.model.institution.InstitutionBase;
 import it.pagopa.selfcare.dashboard.connector.model.user.UserInfo;
 import it.pagopa.selfcare.dashboard.core.InstitutionV2Service;
+import it.pagopa.selfcare.dashboard.core.UserV2Service;
+import it.pagopa.selfcare.dashboard.web.model.InstitutionResource;
 import it.pagopa.selfcare.dashboard.web.model.InstitutionUserResource;
+import it.pagopa.selfcare.dashboard.web.model.mapper.InstitutionResourceMapperImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
@@ -16,6 +21,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import static it.pagopa.selfcare.commons.utils.TestUtils.mockInstance;
 import static java.util.UUID.randomUUID;
 import static org.junit.jupiter.api.Assertions.*;
@@ -24,7 +32,7 @@ import static org.mockito.Mockito.*;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@ContextConfiguration(classes = {InstitutionV2Controller.class})
+@ContextConfiguration(classes = {InstitutionV2Controller.class, InstitutionResourceMapperImpl.class})
 @WebMvcTest(value = {InstitutionV2Controller.class}, excludeAutoConfiguration = SecurityAutoConfiguration.class)
 class InstitutionV2ControllerTest {
 
@@ -36,12 +44,18 @@ class InstitutionV2ControllerTest {
     @Autowired
     protected ObjectMapper objectMapper;
 
+    @MockBean
+    private UserV2Service userServiceMock;
+
     @Autowired
     private InstitutionV2Controller institutionV2Controller;
 
     @MockBean
     private InstitutionV2Service institutionV2ServiceMock;
 
+    /**
+     * Method under test: {@link InstitutionV2Controller#getInstitutionUser(String, String, Authentication)}
+     */
     @Test
     void getInstitutionUser_notNullUser() throws Exception {
         //given
@@ -71,5 +85,41 @@ class InstitutionV2ControllerTest {
                 .getInstitutionUser(institutionId, userId, loggedUserId);
         verifyNoMoreInteractions(institutionV2ServiceMock);
 
+    }
+
+    /**
+     * Method under test: {@link InstitutionV2Controller#getInstitutions(Authentication)}
+     */
+    @Test
+    void getInstitutions_institutionInfoNotNull() throws Exception {
+        // given
+        final String userId = "userId";
+        Authentication authentication = mock(Authentication.class);
+        when(authentication.getPrincipal()).thenReturn(SelfCareUser.builder(userId).build());
+
+        InstitutionBase expectedInstitution = mockInstance(new InstitutionBase());
+        List<InstitutionBase> expectedInstitutionInfos = new ArrayList<>();
+        expectedInstitutionInfos.add(expectedInstitution);
+        when(userServiceMock.getInstitutions(userId)).thenReturn(expectedInstitutionInfos);
+        // when
+        MvcResult result = mvc.perform(MockMvcRequestBuilders
+                        .get(BASE_URL)
+                        .principal(authentication)
+                        .contentType(APPLICATION_JSON_VALUE)
+                        .accept(APPLICATION_JSON_VALUE))
+                .andExpect(status().is2xxSuccessful())
+                .andReturn();
+        // then
+        List<InstitutionResource> resources = objectMapper.readValue(result.getResponse().getContentAsString(),
+                new TypeReference<>() {
+                });
+
+        assertNotNull(resources);
+        assertFalse(resources.isEmpty());
+        assertEquals(resources.get(0).getStatus(), expectedInstitution.getStatus());
+        assertNotNull(resources.get(0).getUserRole());
+        verify(userServiceMock, times(1))
+                .getInstitutions(userId);
+        verifyNoMoreInteractions(userServiceMock);
     }
 }
