+++ web/src/test/java/it/pagopa/selfcare/dashboard/web/controller/UserV2ControllerTest.java
@@ -1,17 +1,13 @@
 package it.pagopa.selfcare.dashboard.web.controller;
 
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import it.pagopa.selfcare.commons.base.security.SelfCareUser;
 import it.pagopa.selfcare.commons.utils.TestUtils;
-import it.pagopa.selfcare.dashboard.connector.model.institution.InstitutionBase;
 import it.pagopa.selfcare.dashboard.connector.model.user.Certification;
 import it.pagopa.selfcare.dashboard.connector.model.user.MutableUserFieldsDto;
 import it.pagopa.selfcare.dashboard.connector.model.user.User;
 import it.pagopa.selfcare.dashboard.connector.model.user.WorkContact;
 import it.pagopa.selfcare.dashboard.core.UserV2Service;
 import it.pagopa.selfcare.dashboard.web.config.WebTestConfig;
-import it.pagopa.selfcare.dashboard.web.model.InstitutionResource;
 import it.pagopa.selfcare.dashboard.web.model.SearchUserDto;
 import it.pagopa.selfcare.dashboard.web.model.UpdateUserDto;
 import it.pagopa.selfcare.dashboard.web.model.mapper.InstitutionResourceMapperImpl;
@@ -28,18 +24,19 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
-import org.springframework.security.core.Authentication;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
 
-import static it.pagopa.selfcare.commons.utils.TestUtils.mockInstance;
 import static org.hamcrest.Matchers.emptyString;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
@@ -64,7 +61,7 @@ class UserV2ControllerTest {
     @Spy
     UserMapperV2 userMapperV2 = new UserMapperV2Impl();
 
-    private static final String BASE_URL = "/v2";
+    private static final String BASE_URL = "/v2/users";
 
     private static final User USER_RESOURCE;
 
@@ -76,39 +73,6 @@ class UserV2ControllerTest {
         workContact.getEmail().setCertification(Certification.SPID);
         workContacts.put("institutionId", workContact);
         USER_RESOURCE.setWorkContacts(workContacts);
-    }
-
-    @Test
-    void getInstitutions_institutionInfoNotNull() throws Exception {
-        // given
-        String userId = "userId";
-        Authentication authentication = mock(Authentication.class);
-        when(authentication.getPrincipal()).thenReturn(SelfCareUser.builder(userId).build());
-
-        InstitutionBase expectedInstitution = mockInstance(new InstitutionBase());
-        List<InstitutionBase> expectedInstitutionInfos = new ArrayList<>();
-        expectedInstitutionInfos.add(expectedInstitution);
-        when(userServiceMock.getInstitutions(userId)).thenReturn(expectedInstitutionInfos);
-        // when
-        MvcResult result = mvc.perform(MockMvcRequestBuilders
-                        .get(BASE_URL + "/institutions")
-                        .principal(authentication)
-                        .contentType(APPLICATION_JSON_VALUE)
-                        .accept(APPLICATION_JSON_VALUE))
-                .andExpect(status().is2xxSuccessful())
-                .andReturn();
-        // then
-        List<InstitutionResource> resources = objectMapper.readValue(result.getResponse().getContentAsString(),
-                new TypeReference<>() {
-                });
-
-        assertNotNull(resources);
-        assertFalse(resources.isEmpty());
-        assertEquals(resources.get(0).getStatus(), expectedInstitution.getStatus());
-        assertNotNull(resources.get(0).getUserRole());
-        verify(userServiceMock, times(1))
-                .getInstitutions(userId);
-        verifyNoMoreInteractions(userServiceMock);
     }
 
     @Test
